fix(sdk): Fix BYTE using int_32 instead of int_8 field in Value proto

+++ sdk/bento/types.py
@@ -6,7 +6,7 @@
 from bento.protos.types_pb2 import Type
 
 # Provide easier access to Proto types
-char = Type(primitive=Type.Primitive.INT32)
+char = Type(primitive=Type.Primitive.BYTE)
 int32 = Type(primitive=Type.Primitive.INT32)
 int64 = Type(primitive=Type.Primitive.INT64)
 float32 = Type(primitive=Type.Primitive.FLOAT32)
